Got integration tests working
