+++ tests/WebApi.IntegrationTests/Common/Fixtures/IntegrationTestBase.cs
@@ -3,41 +3,41 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using SSW.CleanArchitecture.Application.Common.Interfaces;
 using SSW.CleanArchitecture.Infrastructure.Persistence;
+using TUnit.Core.Interfaces;
 
 namespace WebApi.IntegrationTests.Common.Fixtures;
 
 /// <summary>
 /// Integration tests inherit from this to access helper classes
 /// </summary>
-[Collection(TestingDatabaseFixtureCollection.Name)]
-public abstract class IntegrationTestBase : IAsyncLifetime
+// [Collection(TestingDatabaseFixtureCollection.Name)]
+public abstract class IntegrationTestBase : IAsyncInitializer
 {
-    private readonly IServiceScope _scope;
+    private IServiceScope _scope;
 
-    private readonly TestingDatabaseFixture _fixture;
-    protected IMediator Mediator { get; }
+    private WebApplicationFactory _fixture;
+    protected IMediator Mediator { get; private set; }
 
     // TODO: Consider removing this as query results can be cached and cause bad test results
     //       Also, consider encapsulating this and only exposing a `Query` method that internally uses `AsNoTracking()`
     //       see: https://github.com/SSWConsulting/SSW.CleanArchitecture/issues/324
     public IApplicationDbContext Context => _dbContext;
 
-    private readonly ApplicationDbContext _dbContext;
+    private ApplicationDbContext _dbContext;
 
     protected IQueryable<T> GetQueryable<T>() where T : class => _dbContext.Set<T>().AsNoTracking();
 
-    protected IntegrationTestBase(TestingDatabaseFixture fixture, ITestOutputHelper output)
+    protected IntegrationTestBase(WebApplicationFactory fixture)
     {
         _fixture = fixture;
-        _fixture.Factory.Output = output;
-
-        _scope = _fixture.ScopeFactory.CreateScope();
-        Mediator = _scope.ServiceProvider.GetRequiredService<IMediator>();
-        _dbContext = _scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
     }
 
     public async Task InitializeAsync()
     {
+        _scope = _fixture.ScopeFactory.CreateScope();
+        Mediator = _scope.ServiceProvider.GetRequiredService<IMediator>();
+        _dbContext = _scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+
         await _fixture.ResetState();
     }
 
@@ -46,7 +46,7 @@ public abstract class IntegrationTestBase : IAsyncLifetime
         await Context.SaveChangesAsync(cancellationToken);
     }
 
-    protected HttpClient GetAnonymousClient() => _fixture.Factory.AnonymousClient.Value;
+    protected HttpClient GetAnonymousClient() => _fixture.AnonymousClient.Value;
 
     public Task DisposeAsync()
     {
