+++ tests/WebApi.IntegrationTests/Endpoints/Heroes/Commands/UpdateHeroCommandTests.cs
@@ -8,10 +8,11 @@ using WebApi.IntegrationTests.Common.Fixtures;
 
 namespace WebApi.IntegrationTests.Endpoints.Heroes.Commands;
 
-public class UpdateHeroCommandTests(TestingDatabaseFixture fixture, ITestOutputHelper output)
-    : IntegrationTestBase(fixture, output)
+[ClassDataSource<WebApplicationFactory>(Shared = SharedType.PerTestSession)]
+public class UpdateHeroCommandTests(WebApplicationFactory fixture)
+    : IntegrationTestBase(fixture)
 {
-    [Fact]
+    [Test]
     public async Task Command_ShouldUpdateHero()
     {
         // Arrange
@@ -50,7 +51,7 @@ public class UpdateHeroCommandTests(TestingDatabaseFixture fixture, ITestOutputH
         item.UpdatedAt.Should().BeCloseTo(createdTimeStamp, TimeSpan.FromSeconds(10));
     }
 
-    [Fact]
+    [Test]
     public async Task Command_WhenHeroDoesNotExist_ShouldReturnNotFound()
     {
         // Arrange
