+++ tests/WebApi.IntegrationTests/Endpoints/Teams/Commands/CompleteMissionCommandTests.cs
@@ -10,7 +10,7 @@ namespace WebApi.IntegrationTests.Endpoints.Teams.Commands;
 public class CompleteMissionCommandTests(TestingDatabaseFixture fixture, ITestOutputHelper output)
     : IntegrationTestBase(fixture, output)
 {
-    [Fact]
+    [Test]
     public async Task Command_ShouldCompleteMission()
     {
         // Arrange
