+++ tests/WebApi.IntegrationTests/Usings.cs
@@ -1,3 +1,3 @@
-global using Xunit;
-global using Xunit.Abstractions;
+// global using Xunit;
+// global using Xunit.Abstractions;
 global using FluentAssertions;
